Move some code to runtime jar
